Move Navigation item component into docs-ui
