+++ packages/docs-ui/src/index.ts
@@ -1,2 +1,4 @@
 export { LinkableHeading } from './components/LinkableHeading/LinkableHeading';
 export { MenuButton } from './components/MenuButton/MenuButton';
+export { NavigationItem } from './components/NavigationSidebar/NavigationItem';
+export type { SubNavigationItem } from './components/NavigationSidebar/NavigationItem';
